+++ auth0/resource_auth0_connection_test.go
@@ -805,29 +805,42 @@ resource "auth0_connection" "email" {
 }
 
 func TestAccConnection_Salesforce(t *testing.T) {
-
 	rand := random.String(6)
-
 	resource.Test(t, resource.TestCase{
 		ProviderFactories: testAccProviderFactories,
 		Steps: []resource.TestStep{
 			{
+				// language=HCL
 				Config: random.Template(`
-
-resource "auth0_connection" "salesforce_community" {
+resource "auth0_connection" "salesforce" {
 	name = "Acceptance-Test-Salesforce-Connection-{{.random}}"
 	is_domain_connection = false
 
-	options {
+	salesforce {		
 		client_id = "client-id"
 		client_secret = "client-secret"
-		community_base_url = "https://salesforce.example.com"
+		set_user_root_attributes = "on_each_login"
+		non_persistent_attrs = ["foo", "bar"]
+		scopes = [ "email", "profile", "gmail", "youtube" ]
 	}
 }
 `, rand),
 				Check: resource.ComposeAggregateTestCheckFunc(
-					random.TestCheckResourceAttr("auth0_connection.salesforce_community", "name", "Acceptance-Test-Salesforce-Connection-{{.random}}", rand),
-					resource.TestCheckResourceAttr("auth0_connection.salesforce_community", "options.0.community_base_url", "https://salesforce.example.com"),
+					random.TestCheckResourceAttr("auth0_connection.salesforce", "name", "Acceptance-Test-Salesforce-Connection-{{.random}}", rand),
+					resource.TestCheckResourceAttr("auth0_connection.salesforce", "salesforce.0.client_id",
+						"client-id"),
+					resource.TestCheckResourceAttr("auth0_connection.salesforce", "salesforce.0.client_secret",
+						"client-secret"),
+					resource.TestCheckResourceAttr("auth0_connection.salesforce", "salesforce.0.scopes.#", "4"),
+					resource.TestCheckResourceAttr("auth0_connection.salesforce", "salesforce.0.scopes.0", "email"),
+					resource.TestCheckResourceAttr("auth0_connection.salesforce", "salesforce.0.scopes.2", "profile"),
+					resource.TestCheckResourceAttr("auth0_connection.salesforce", "salesforce.0.scopes.1", "gmail"),
+					resource.TestCheckResourceAttr("auth0_connection.salesforce", "salesforce.0.scopes.3", "youtube"),
+					resource.TestCheckResourceAttr("auth0_connection.salesforce", "salesforce.0.set_user_root_attributes", "on_each_login"),
+					resource.TestCheckResourceAttr("auth0_connection.salesforce",
+						"salesforce.0.non_persistent_attrs.0", "bar"),
+					resource.TestCheckResourceAttr("auth0_connection.salesforce",
+						"salesforce.0.non_persistent_attrs.1", "foo"),
 				),
 			},
 		},
@@ -852,7 +865,6 @@ resource "auth0_connection" "google_oauth2" {
 		allowed_audiences = [ "example.com", "api.example.com" ]		
 		set_user_root_attributes = "on_each_login"
 		non_persistent_attrs = ["foo", "bar"]
-		scopes = [ "email", "profile", "gmail", "youtube" ]
 	}
 }
 `, rand),
@@ -865,11 +877,6 @@ resource "auth0_connection" "google_oauth2" {
 						"example.com"),
 					resource.TestCheckResourceAttr("auth0_connection.google_oauth2", "google_oauth2.0.allowed_audiences.0",
 						"api.example.com"),
-					resource.TestCheckResourceAttr("auth0_connection.google_oauth2", "google_oauth2.0.scopes.#", "4"),
-					resource.TestCheckResourceAttr("auth0_connection.google_oauth2", "google_oauth2.0.scopes.0", "email"),
-					resource.TestCheckResourceAttr("auth0_connection.google_oauth2", "google_oauth2.0.scopes.2", "profile"),
-					resource.TestCheckResourceAttr("auth0_connection.google_oauth2", "google_oauth2.0.scopes.1", "gmail"),
-					resource.TestCheckResourceAttr("auth0_connection.google_oauth2", "google_oauth2.0.scopes.3", "youtube"),
 					resource.TestCheckResourceAttr("auth0_connection.google_oauth2", "google_oauth2.0.set_user_root_attributes", "on_each_login"),
 					resource.TestCheckResourceAttr("auth0_connection.google_oauth2",
 						"google_oauth2.0.non_persistent_attrs.0", "bar"),